Move Servlet-specific observation classes to dedicated package
