+++ spring-web/src/main/java/org/springframework/web/observation/servlet/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Instrumentation for {@link io.micrometer.observation.Observation observing} Servlet web applications.
+ */
+@NonNullApi
+@NonNullFields
+package org.springframework.web.observation.servlet;
+
+import org.springframework.lang.NonNullApi;
+import org.springframework.lang.NonNullFields;
