+++ spring-web/src/test/java/org/springframework/web/observation/HttpRequestsObservationFilterTests.java
@@ -23,6 +23,8 @@ import jakarta.servlet.ServletException;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.http.HttpMethod;
+import org.springframework.web.observation.servlet.HttpRequestsObservationContext;
+import org.springframework.web.observation.servlet.HttpRequestsObservationFilter;
 import org.springframework.web.testfixture.servlet.MockFilterChain;
 import org.springframework.web.testfixture.servlet.MockHttpServletRequest;
 import org.springframework.web.testfixture.servlet.MockHttpServletResponse;
